refactor: move access control allow/denylist to own file
